Перенес функцию фильтра в файл компонента фильтра.
